% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPatternInFiles.R
\name{findPatternInFiles}
\alias{findPatternInFiles}
\title{Find a pattern in files}
\usage{
findPatternInFiles(ext, pattern, depth = NULL, wholeword = FALSE)
}
\arguments{
\item{ext}{file extension}

\item{pattern}{pattern to search}

\item{depth}{depth of the search; \code{0} for the current directory,
\code{1} for the current directory and its subdirectories, etc, and
\code{NULL} for no limit}

\item{wholeword}{logical, whether to search the pattern as a whole}
}
\description{
Find pattern in files and display the results in RStudio viewer.
}
